+++ bin/utop.ml
@@ -31,6 +31,7 @@ let term =
         let* setup = Import.Main.setup () in
         Build_system.run_exn (fun () ->
             let open Memo.Build.O in
+            let* setup = setup in
             let context = Import.Main.find_context_exn setup ~name:ctx_name in
             let sctx = Import.Main.find_scontext_exn setup ~name:ctx_name in
             let utop_target =