+++ src/dune_engine/package.mli
@@ -154,6 +154,8 @@ val dir : t -> Path.Source.t
 
 val file : dir:Path.t -> name:Name.t -> Path.t
 
+val encode : t Dune_lang.Encoder.t
+
 val decode : dir:Path.Source.t -> t Dune_lang.Decoder.t
 
 val opam_file : t -> Path.Source.t