+++ src/dune_engine/response_file.ml
@@ -6,7 +6,8 @@ type t =
 
 (* This mutable table is safe under the assumption that a program path always
    points to the binary with the same version. While the assumption seems likely
-   to hold, it would be better to avoid the need for it to simplify reasoning. *)
+   to hold, it would be better to avoid the need for it to simplify
+   reasoning. *)
 let registry = Table.create (module Path) 128
 
 let get ~prog = Option.value (Table.find registry prog) ~default:Not_supported