+++ src/dune_rules/compilation_context.ml
@@ -128,7 +128,8 @@ let create ~super_context ~scope ~expander ~obj_dir ~modules ~flags
     (* With sandboxing, there are a few build errors in ocaml platform 1162238ae
        like: File "ocaml_modules/ocamlgraph/src/pack.ml", line 1: Error: The
        implementation ocaml_modules/ocamlgraph/src/pack.ml does not match the
-       interface ocaml_modules/ocamlgraph/src/.graph.objs/byte/graph__Pack.cmi: *)
+       interface
+       ocaml_modules/ocamlgraph/src/.graph.objs/byte/graph__Pack.cmi: *)
     Sandbox_config.no_sandboxing
   in
   let modes =