+++ src/dune_rules/foreign_rules.ml
@@ -9,7 +9,8 @@ module Source_tree_map_reduce =
     end))
 
 (* Compute command line flags for the [include_dirs] field of [Foreign.Stubs.t]
-   and track all files in specified directories as [Hidden_deps] dependencies. *)
+   and track all files in specified directories as [Hidden_deps]
+   dependencies. *)
 let include_dir_flags ~expander ~dir (stubs : Foreign.Stubs.t) =
   let scope = Expander.scope expander in
   let lib_dir loc lib_name =
@@ -37,7 +38,8 @@ let include_dir_flags ~expander ~dir (stubs : Foreign.Stubs.t) =
                match Path.extract_build_context_dir include_dir with
                | None ->
                  (* This branch corresponds to an external directory. The
-                    current implementation tracks its contents NON-recursively. *)
+                    current implementation tracks its contents
+                    NON-recursively. *)
                  (* TODO: Track the contents recursively. One way to implement
                     this is to change [Build_system.Loaded.Non_build] so that it
                     contains not only files but also directories and traverse